Jira: DAISY-30 Baremetal Deployment on zte-pod2
do not merge till the test is pass

+++ ci/deploy/deploy.sh
@@ -35,6 +35,7 @@ compute_node_size=`echo $parameter_from_deploy | cut -d " " -f 3`
 daisy_passwd=`echo $parameter_from_deploy | cut -d " " -f 4`
 daisy_ip=`echo $parameter_from_deploy | cut -d " " -f 5`
 daisy_gateway=`echo $parameter_from_deploy | cut -d " " -f 6`
+deploy_env=`echo $parameter_from_deploy | cut -d " " -f 7`
 
 VM_STORAGE=/home/qemu/vms
 test -d ${VM_STORAGE} || mkdir -p ${VM_STORAGE}
@@ -104,7 +105,9 @@ function clean_up
 }
 
 echo "=====clean up all node and network======"
-clean_up all_in_one daisy2
+if [ $deploy_env == "virtual" ];then
+    clean_up all_in_one daisy2
+fi
 clean_up daisy daisy1
 if [ -f $WORKDIR/daisy/centos7.qcow2 ]; then
     rm -rf $WORKDIR/daisy/centos7.qcow2
@@ -138,24 +141,41 @@ execute_on_jumpserver $daisy_ip "echo -e '[libvirt]\nvirt_type=qemu' > /etc/koll
 echo "===prepare cluster and pxe==="
 execute_on_jumpserver $daisy_ip "python ${REMOTE_SPACE}/deploy/tempest.py --dha $DHA --network $NETWORK --cluster 'yes'"
 
-echo "=====create all-in-one node======"
-qemu-img create -f qcow2 ${VM_STORAGE}/all_in_one.qcow2 200G
-create_node $net_daisy2 daisy2 $pod_all_in_one all_in_one
-sleep 20
+echo "=====create and find node======"
+if [ $deploy_env == "virtual" ];then
+    qemu-img create -f qcow2 ${VM_STORAGE}/all_in_one.qcow2 200G
+    create_node $net_daisy2 daisy2 $pod_all_in_one all_in_one
+    sleep 20
+else
+    ipmitool -I lanplus -H 192.168.1.106 -U zteroot -P superuser -R 1 chassis bootdev pxe
+    ipmitool -I lanplus -H 192.168.1.106 -U zteroot -P superuser -R 1 chassis  power reset
+    ipmitool -I lanplus -H 192.168.1.107 -U zteroot -P superuser -R 1 chassis bootdev pxe
+    ipmitool -I lanplus -H 192.168.1.107 -U zteroot -P superuser -R 1 chassis  power reset
+    ipmitool -I lanplus -H 192.168.1.108 -U zteroot -P superuser -R 1 chassis bootdev pxe
+    ipmitool -I lanplus -H 192.168.1.108 -U zteroot -P superuser -R 1 chassis  power reset
+    ipmitool -I lanplus -H 192.168.1.109 -U zteroot -P superuser -R 1 chassis bootdev pxe
+    ipmitool -I lanplus -H 192.168.1.109 -U zteroot -P superuser -R 1 chassis  power reset
+    ipmitool -I lanplus -H 192.168.1.110 -U zteroot -P superuser -R 1 chassis bootdev pxe
+    ipmitool -I lanplus -H 192.168.1.110 -U zteroot -P superuser -R 1 chassis  power reset
+fi
 
 echo "======prepare host and pxe==========="
 execute_on_jumpserver $daisy_ip "python ${REMOTE_SPACE}/deploy/tempest.py  --dha $DHA --network $NETWORK --host 'yes'"
 
 echo "======daisy deploy os and openstack==========="
-virsh destroy all_in_one
-virsh start all_in_one
+if [ $deploy_env == "virtual" ];then
+    virsh destroy all_in_one
+    virsh start all_in_one
+fi
 
 echo "===========check install progress==========="
 execute_on_jumpserver $daisy_ip "systemctl restart daisy-api"
 execute_on_jumpserver $daisy_ip "systemctl restart daisy-registry"
 execute_on_jumpserver $daisy_ip "${REMOTE_SPACE}/deploy/check_os_progress.sh"
 sleep 10
-virsh reboot all_in_one
+if [ $deploy_env == "virtual" ];then
+    virsh reboot all_in_one
+fi
 execute_on_jumpserver $daisy_ip "${REMOTE_SPACE}/deploy/check_openstack_progress.sh"
 
 exit 0
